+++ crates/flipperzero/src/prelude.rs
@@ -1,0 +1,4 @@
+//! Flipper Zero Rust Prelude
+
+pub use crate::furi::string::FuriString;
+pub use crate::{format, print, println};